InterzonalAirflow.sim and n50 as default
Noticed that no default was provided anymore for n50 zone, in the current implementation the easiest way to solve this is to provide the default as the start value.

This works but will promt a warning during translation pointing out to the user that not really a value was provided.