Wind speed modifier according to ASHRAE 1993/CONTAM
SimInfomanager calculates one Cs for the building surfaces (regardless of effective height).

Default is suburban and hroof equal to href (no wind velocity profile correction)

Same calculation also available in 'OutsideAir' where it expects the height of the boundary as input. Not used in implementation for now as heights of zones and surfaces not known yet. Once this is done, the Cs in sim can be eliminated.